Even more tests
This fix should finish rounding out the majority of the spec suite.
There are about 97 tests that are current stubbed out inbetween
the spec and the site/tests dir, mostly dealing with:

- server side attributes
- server side status methods
- ruby client interface operations (and bot operations)
- various random / small operations / edge cases

There may be a few other things in the codebase for which tests
haven't been implemented or stubbed out yet but I consider those
to be bugs to be resolved as they are discovered.

With this commit all of the required / critical / core functionality
of the app and clients should have unit tests.