Support Omega Events in Users subsystem
Adds a RegisteredUser event and mechanism
to subscribe to / unsubscribe to events.

Slight changes to Omega Event to add persistence
flag to EventHandler. If set handler will not
be deleted from registry upon being run.

Specs for everything are implemented except
for users event rjr methods which are stubbed out