UI: utilize recursive/children flags to optimize lookup (WIP)
Cosmos entities are now lazy loaded in the UI so that large
universes are processed as quickly as possible.

Still some kinks to work out as altering the load order is
resulting in some breakage due to data not being available
and the necessary callbacks not being invoked.

Need to look into resolving these issue and perhaps
simplifying the js client loading workflow