UI: use cosmos::interconnects to retrieve all system interconns
This utilizes the newly added cosmos::interconnects method
in the UI to load all the interconnects in a galaxy even if
all the systems have not been loaded.

There is a regression w/ this commit where clicking on a system
in a galaxy breaks due to children not being loaded. Clicking on
a system in the locations menu still works.

The current server synchronization phase in the js client is
quite convoluted. Cleaning it up is currently a high priority
item