UI: flush out initial missiles / heavy weapons class gfx
Not the prettiest implementation as a few hacky callbacks had
to be added to canvas to handle a few edge cases. Most likely
will be looking to incorporate improvements into some high-level
js entity & component level refactoring in the near future.

Tests are also (more) broken post this, will incorporate fixes
with updates